feat: use current networkId in notification
this was missed in 7ab9ae34479dd098edf609b83ac78484c7b564fa

in the notification that links people to NEAR Explorer, we need to link
to the version of explorer for their selected network. This way, if
someone runs their new app on betanet, the links in the notification
don't lead to 404 pages.