chore: work around ignored npmignore
`npm publish` is supposed to ignore `.npmignore`, but it doesn't work
reliably: https://stackoverflow.com/q/63046027/249801

Given that release-it requires a clean working directory and this
prerelease script dirties the working directory, this will result in
some annoyance. However, that annoyance is better than accidentally
publishing hundreds of megabytes of Rust build artifacts.