refactor(utils.js): mv out of common; rm onSubmit
onSubmit never felt quite right as a shared function, given that it
manipulates the DOM

Additionally, the vue cli added in #452 does not work with `utils.js`
being in a symlink and `config.js` being in the project directory, so
they both need to live in the project directory.

Ideally, near-api-js could be simplified to a point that the helpers in
utils.js are superfluous. That will be a better way to get rid of the
duplication in each template, rather than using a symlink.