build: up all deps; use more conservative versions
We don't include `yarn.lock` or `package-lock.json` in the files copied
into the project, because we don't want to enforce project authors to
use one or the other. But this means that the dependency versions used
locally within this project and its tests can differ significantly from
the versions installed by a project that uses create-near-app. In the
case of jest, I found that running `./index.js` to create a test React
project ended up with `jest@26.1.0`, whereas a project created with `npx
create-near-app` ended up with `jest@26.6.2`

A good remediation is to use tilde version numbering (`~26.6.0`), which
specifies that patch-level variation is tolerable (`26.6.100` is fine,
but `26.7.0` is not), rather than caret version numbering (`^26.6.0`),
which specifies that minor-level variation is also fine (`26.100.0`
would be fine in this case).

In addition to using tilde versioning, I also ran `yarn upgrade
--latest` in all template directories except Angular to give us a fresh
start on using latest dependencies. I noticed that jest was still set to
`26.0.1` for the React project. Very old at this point!

Why not Angular? Because it kept failing on Travis for macOS, even
though the tests worked locally. Ok then, let's let Dependabot deal with
this one, one dependency at a time, to suss out the culprit.

build: angular-versions-match-master